Colours and palettes bas

+++ src/dataIO/PovExport.java
@@ -182,7 +182,7 @@ public class PovExport implements GraphicalExporter
 		}
 		else
 		{
-			double[] pigmentArray = (double[]) pigment;
+			float[] pigmentArray = (float[]) pigment;
 			String rgbStatement = "color rgb < " + pigmentArray[0] + ", " + 
 				pigmentArray[1] + ", " + pigmentArray[2] + " >";
 			return rgbStatement;
